+++ libs/designer/src/lib/ui/panel/panelroot.tsx
@@ -2,8 +2,8 @@ import constants from '../../common/constants';
 import type { AppDispatch, RootState } from '../../core';
 import { deleteOperation } from '../../core/actions/bjsworkflow/delete';
 import { useIsXrmConnectionReferenceMode, useMonitoringView, useReadOnly } from '../../core/state/designerOptions/designerOptionsSelectors';
-import { updateParameterValidation } from '../../core/state/operation/operationMetadataSlice';
-import { useParameterValidationErrors } from '../../core/state/operation/operationSelector';
+import { ErrorLevel, updateParameterValidation } from '../../core/state/operation/operationMetadataSlice';
+import { useOperationErrorInfo, useParameterValidationErrors } from '../../core/state/operation/operationSelector';
 import {
   useCurrentPanelModePanelMode,
   useIsPanelCollapsed,
@@ -80,46 +80,45 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
   let showCommentBox = !isNullOrUndefined(comment);
   const hasSchema = useHasSchema(operationInfo?.connectorId, operationInfo?.operationId);
   const isXrmConnectionReferenceMode = useIsXrmConnectionReferenceMode();
+  const errorInfo = useOperationErrorInfo(selectedNode);
+
+  const selectConnectionTabTitle = isXrmConnectionReferenceMode
+    ? intl.formatMessage({
+        defaultMessage: 'Select Connection Reference',
+        description: 'Title for the select connection reference tab',
+      })
+    : intl.formatMessage({
+        defaultMessage: 'Select Connection',
+        description: 'Title for the select connection tab',
+      });
+  const createConnectionTabTitle = isXrmConnectionReferenceMode
+    ? intl.formatMessage({
+        defaultMessage: 'Create Connection Reference',
+        description: 'Title for the create connection reference tab',
+      })
+    : intl.formatMessage({
+        defaultMessage: 'Create Connection',
+        description: 'Title for the create connection tab',
+      });
 
   useEffect(() => {
-    const selectConnectionTabTitle = isXrmConnectionReferenceMode
-      ? intl.formatMessage({
-          defaultMessage: 'Select Connection Reference',
-          description: 'Title for the select connection reference tab',
-        })
-      : intl.formatMessage({
-          defaultMessage: 'Select Connection',
-          description: 'Title for the select connection tab',
-        });
-    const selectConnectionTab = getSelectConnectionTab(selectConnectionTabTitle);
-
-    const createConnectionTabTitle = isXrmConnectionReferenceMode
-      ? intl.formatMessage({
-          defaultMessage: 'Create Connection Reference',
-          description: 'Title for the create connection reference tab',
-        })
-      : intl.formatMessage({
-          defaultMessage: 'Create Connection',
-          description: 'Title for the create connection tab',
-        });
-    const createConnectionTab = getCreateConnectionTab(createConnectionTabTitle);
-
-    const tabs = [
-      monitoringTab,
-      parametersTab,
-      settingsTab,
-      codeViewTab,
-      testingTab,
-      createConnectionTab,
-      selectConnectionTab,
-      aboutTab,
-      loadingTab,
-    ];
+    const tabs = [monitoringTab, parametersTab, settingsTab, codeViewTab, testingTab, aboutTab, loadingTab];
     if (process.env.NODE_ENV !== 'production') {
       tabs.push(scratchTab);
     }
     dispatch(registerPanelTabs(tabs));
-  }, [dispatch, isXrmConnectionReferenceMode, intl]);
+    dispatch(clearPanel());
+  }, [dispatch]);
+
+  useEffect(() => {
+    const createConnectionTab = getCreateConnectionTab(createConnectionTabTitle);
+    dispatch(registerPanelTabs([createConnectionTab]));
+  }, [dispatch, createConnectionTabTitle]);
+
+  useEffect(() => {
+    const selectConnectionTab = getSelectConnectionTab(selectConnectionTabTitle);
+    dispatch(registerPanelTabs([selectConnectionTab]));
+  }, [dispatch, selectConnectionTabTitle]);
 
   useEffect(() => {
     dispatch(
@@ -177,9 +176,9 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
 
   const parameterValidationErrors = useParameterValidationErrors(selectedNode);
   useEffect(() => {
-    const hasErrors = parameterValidationErrors?.length > 0;
+    const hasErrors = parameterValidationErrors?.length > 0 || errorInfo?.level === ErrorLevel.Connection;
     dispatch(setTabError({ tabName: 'parameters', hasErrors, nodeId: selectedNode }));
-  }, [dispatch, parameterValidationErrors?.length, selectedNode]);
+  }, [dispatch, errorInfo?.level, parameterValidationErrors?.length, selectedNode]);
 
   useEffect(() => {
     collapsed ? setWidth(PanelSize.Auto) : setWidth(PanelSize.Medium);
@@ -268,6 +267,10 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
     dispatch(replaceId({ originalId: selectedNode, newId }));
   };
 
+  const onCommentChange = (newDescription?: string) => {
+    dispatch(setNodeDescription({ nodeId: selectedNode, description: newDescription }));
+  };
+
   const handleDelete = (): void => {
     dispatch(deleteOperation({ nodeId: selectedNode, isTrigger: isTriggerNode }));
     // TODO: 12798935 Analytics (event logging)
@@ -307,7 +310,8 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
       cardIcon={iconUri}
       comment={comment}
       noNodeSelected={!selectedNode}
-      isError={opQuery?.isError}
+      isError={errorInfo?.level === ErrorLevel.Critical || opQuery?.isError}
+      errorMessage={errorInfo?.message}
       isLoading={isLoading}
       panelScope={PanelScope.CardLevel}
       panelHeaderControlType={getPanelHeaderControlType() ? PanelHeaderControlType.DISMISS_BUTTON : PanelHeaderControlType.MENU}
@@ -334,9 +338,7 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
         togglePanel();
       }}
       trackEvent={handleTrackEvent}
-      onCommentChange={(value) => {
-        dispatch(setNodeDescription({ nodeId: selectedNode, description: value }));
-      }}
+      onCommentChange={onCommentChange}
       title={selectedNodeDisplayName}
       onTitleChange={onTitleChange}
     />